+++ sync-for-expenses/src/codatsyncexpenses/models/shared/customer.py
@@ -11,6 +11,7 @@ from typing import Optional
 
 @dataclasses.dataclass
 class Customer:
+    r"""Success"""
     id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('id'), 'exclude': lambda f: f is None }})
     r"""id of the customer for all income related activities to be associated to."""