+++ sync-for-expenses/src/codatsyncexpenses/sdkconfiguration.py
@@ -17,8 +17,8 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = 'prealpha'
-    sdk_version: str = '0.34.0'
-    gen_version: str = '2.84.3'
+    sdk_version: str = '0.35.0'
+    gen_version: str = '2.91.4'
 
     def get_server_details(self) -> tuple[str, dict[str, str]]:
         if self.server_url: