+++ sync-for-expenses/src/codatsyncexpenses/transaction_status.py
@@ -13,7 +13,7 @@ class TransactionStatus:
         self.sdk_configuration = sdk_config
         
     
-    def get_sync_transaction(self, request: operations.GetSyncTransactionRequest, retries: Optional[utils.RetryConfig] = None) -> operations.GetSyncTransactionResponse:
+    def get(self, request: operations.GetSyncTransactionRequest, retries: Optional[utils.RetryConfig] = None) -> operations.GetSyncTransactionResponse:
         r"""Get Sync Transaction
         Gets the status of a transaction for a sync
         """
@@ -44,8 +44,8 @@ class TransactionStatus:
         
         if http_res.status_code == 200:
             if utils.match_content_type(content_type, 'application/json'):
-                out = utils.unmarshal_json(http_res.text, Optional[list[shared.TransactionMetadata]])
-                res.transaction_metadata = out
+                out = utils.unmarshal_json(http_res.text, Optional[shared.Transaction])
+                res.transaction = out
             else:
                 raise errors.SDKError(f'unknown content-type received: {content_type}', http_res.status_code, http_res.text, http_res)
         elif http_res.status_code in [401, 404, 429]:
@@ -58,9 +58,9 @@ class TransactionStatus:
         return res
 
     
-    def list_sync_transactions(self, request: operations.ListSyncTransactionsRequest, retries: Optional[utils.RetryConfig] = None) -> operations.ListSyncTransactionsResponse:
-        r"""Get Sync transactions
-        Get's the transactions and status for a sync
+    def list(self, request: operations.ListSyncTransactionsRequest, retries: Optional[utils.RetryConfig] = None) -> operations.ListSyncTransactionsResponse:
+        r"""List sync transactions
+        Gets the transactions and status for a sync
         """
         base_url = utils.template_url(*self.sdk_configuration.get_server_details())
         
@@ -90,8 +90,8 @@ class TransactionStatus:
         
         if http_res.status_code == 200:
             if utils.match_content_type(content_type, 'application/json'):
-                out = utils.unmarshal_json(http_res.text, Optional[shared.TransactionMetadataList])
-                res.transaction_metadata_list = out
+                out = utils.unmarshal_json(http_res.text, Optional[shared.Transactions])
+                res.transactions = out
             else:
                 raise errors.SDKError(f'unknown content-type received: {content_type}', http_res.status_code, http_res.text, http_res)
         elif http_res.status_code in [401, 404, 429]: